pyproject.toml
[project]
name = "rivus-ai-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.4.9",
    "fastapi>=0.115.0",
    "httpx>=0.28.1",
    "langchain>=1.2.11",
    "langchain-openai>=1.1.11",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.22",
    "uvicorn[standard]>=0.34.0",
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["src"]
